Remove unused code on `packages/core-typings`
