+++ packages/core-typings/src/index.ts
@@ -75,7 +75,7 @@ export * from './IOembed';
 export * from './IEmailInbox';
 export * from './ILoginServiceConfiguration';
 export * from './ISocketConnection';
-export * from './IMethodThisType';
+export * from './IMethodConnection';
 export * from './IPassword';
 export * from './IBaseData';
 export * from './IOAuthApps';
@@ -84,8 +84,6 @@ export * from './IOAuthAccessToken';
 export * from './IOAuthRefreshToken';
 export * from './ISmarshHistory';
 export * from './IReport';
-
-// export * from './IMethodConnection';
 
 export * from './ldap';
 // TODO: move to separated package
